[layout] Convert MOVSS/MOVSD X86 instructions to not rematerializable.
In X86 MOVSDrm_alt is rematerializable, which means that it doesn't have to be spilled in memory.
Essentially, X86 uses rip-relative addressing to get the constant values of the global variables,
directly from the stack, whereas AArch64 loads the constant to a GPR, and then copies the GPR to the FPR.
So, we always treat MOVSS/MOVSD as not rematerializable, so X86 will be forced to also spill in the stack.