[layout] Treat LEA64 x86 instructions as cheap as a move.
When we are accessing references inside a loop, X86 will usually use LEA
instructions that are hoisted outside of the loop. This causes greater
register pressure. The main problem, however, is that AArch64 will use
simple arithmetic instructions, which are usually rematerialized instead
of hoisted. To fix that difference, we set the `isAsCheapAsAMove` flag
for LEA64r, so that `simple-register-coalescing` will also rematerialize LEAs.
Moreover, this makes the flag `sink-insts-to-avoid-spills` unnecessary.