[layout] Immediates in instructions have the same size as in AArch64.
We encode in instructions only immediates that can be represented
using 12 bits, with an optional bit to declare that the immediate
is a value shifted by 12. Otherwise, a register is used to hold larger
immediates. This is how AArch64 represents immediates.
We support this currently for instructions that are using constant
values in the DAG.