[layout] Add SHF_MERGE flag for section with known entry size.
The SHF_MERGE flag is expected when the entry size for the section's
entries is known. This is not enforced as a general error, but only as
an assertion.
Thus, it is unclear how this information trickles down to target
assembly and what linkers might do during section merging if any of
these two fields are missing.