[layout] Do not move directly 32-bit immediates to memory in X86.
In the case of AArch64, these immediates need to be loaded into a
register first. Sometimes, a CSR will be used, which breaks the
unified address space with X86, which can move directly immediates
to memory. Add a flag to deactivate this X86 capability.

Addresses: https://github.com/systems-nuts/UnASL/issues/104