[layout] Add X86 flag for following AArch64 constant cost model.
AArch64 and X86 follow a different strategy for materializing constants,
e.g., whether to encode the full constant, or encode a smaller one and
add the rest of the value as a separate instruction.
We want to align both architectures in this respect.

Addresses: https://github.com/systems-nuts/UnASL/issues/115