[layout] Handle architecture-specific stack slots in stackmaps.
So far, architecture-specific values that are stored in the stack
were not included in the stackmap section of the ELF binary.
For example, when the value of a global constant was spilled,
we weren't including this liveness information in the stackmaps.
Thus, our tools could not detect whether this live value would end
up in the same stack location.

This patch introduces support for these stack slots and also handles
X86 values generated by rip-relative memory-to-register operations,
with `MOVSDrm_alt`.

Addresses: https://github.com/systems-nuts/UnASL/issues/87
Addresses: https://github.com/systems-nuts/UnASL/issues/88