[layout] Fix prolog setup in the case of odd number of CSRs in AArch64.
When spilling an emergency slot and the `CombineSPBump` is true,
we need to calculate the local stack size differently depending on
the parity of CSRs. If the number of CSRs is odd, then we need
to explicitly include the emergency slot in the local stack size,
because it is hidden by the alignment of `PrologueSaveSize`.

Addresses: https://github.com/systems-nuts/UnASL/issues/184