[layout] Add flag to enable the SimplifyCFG pass in X86
By default, SimplifyCFG does not happen on the X86 target. On AArch64,
it occurs right after the AtomicExpand pass. We add a flag to enable this
for X86 also.

However, this breaks test cases where we had disabled
hoisting during lowering in AArch64, e.g., which was fixed by:

https://github.com/blackgeorge-boom/llvm-project/pull/15

Now, instead of disabling hoisting in AArch64, it seems simpler to
enable the SimplifyCFG in X86 (which is the reason why there was the
difference in the first place in the case of hoisting in AArch64).
So, by enabling the SimplifyCFG pass, we fix both of these issues:

* https://github.com/systems-nuts/unifico/issues/73 (older issue)
* https://github.com/systems-nuts/unifico/issues/241 (current issue)

For now, we can deactivate the AArch64 flag `attr=+disable-hoist-in-lowering`,
and only keep the `-x86-enable-simplify-cfg` flag for both of the above
issues. We will revisit this in the future, if need be.