[layout] Copy WZR only to temporary registers
The way AArch64 propagates the zero value is by assigning the zero
register to another register, which is, potentially, copied further.
Since with X86 we implement the equivalent functionality with `MOV32r0`,
which we only copy to temporary registers, we need to make sure that in
AArch64 the zero register and its subsequent copies are only allowed
to temp registers as well.

Addresses: https://github.com/systems-nuts/unifico/issues/256