[layout] Fix order of checking VRs after arch-specific value creation
When we are adding arch-specific stack slots, we need to use
`sanitizeVRegs` after we create a value, not before. Otherwise, we will
not check correctly if a virtual register has not been handled by the
stackmap and when we will try to lower the register as part of the
arch-specific value, we will fail.

Addresses: https://github.com/systems-nuts/unifico/issues/275