[layout] AArch64: use the two-address instruction format for ADD/SUB
It seems that in many cases, the only way to control the register
allocation is to impose the same two-address instruction format between
the architectures. Otherwise, the different live ranges lead to
different allocation of CSRs. Also, need to make these instructions
commutable, in order to allow other smart tricks of the
`twoaddressinstruction` pass (e.g., to enable more coalescing).

Maybe in the future we'll find a better way.

Addresses: https://github.com/systems-nuts/unifico/issues/283