[layout] Emulate three-address conversion in AArch64
Since we are already emulating the conversion of AArch64's ADD/SUB-rr
instructions to a two-address version, we need to also emulate the conversion
of those back to their original three-address format. This mimics what X86
is doing, by converting two-address additions to a three-address LEA.

Potentially, we need to expand this to other operations, like shifts.