[layout] Fix bug when coalescing temp regs to a subreg.
When we are coalescing a temp reg class, instead of blindly propagating
the temporary class, we need to check whether the destination is a
subregister. In this case, we may need to expand the temp register
class, so that it fits the size of the destination class being subregistered.

Addresses: https://github.com/systems-nuts/unifico/issues/286